Move ServerClusterInterfaceRegistry to app/server-cluster to make it easier to reuse
