+++ src/app/clusters/network-commissioning/CodegenInstance.h
@@ -19,7 +19,7 @@
 #include "NetworkCommissioningCluster.h"
 
 #include <data-model-providers/codegen/CodegenDataModelProvider.h>
-#include <data-model-providers/codegen/ServerClusterInterfaceRegistry.h>
+#include <app/server-cluster/ServerClusterInterfaceRegistry.h>
 
 namespace chip {
 namespace app {
