+++ src/app/server-cluster/BUILD.gn
@@ -38,3 +38,17 @@ source_set("server-cluster") {
     "${chip_root}/src/lib/support",
   ]
 }
+
+source_set("registry") {
+  sources = [
+    "ServerClusterInterfaceRegistry.cpp",
+    "ServerClusterInterfaceRegistry.h",
+  ]
+
+  public_deps = [
+    "${chip_root}/src/app:paths",
+    "${chip_root}/src/app/server-cluster",
+    "${chip_root}/src/lib/core:types",
+    "${chip_root}/src/lib/support",
+  ]
+}
