+++ src/app/server-cluster/tests/BUILD.gn
@@ -21,6 +21,7 @@ chip_test_suite("tests") {
   test_sources = [
     "TestAttributeListBuilder.cpp",
     "TestDefaultServerCluster.cpp",
+    "TestServerClusterInterfaceRegistry.cpp",
   ]
 
   cflags = [ "-Wconversion" ]
@@ -29,6 +30,7 @@ chip_test_suite("tests") {
     "${chip_root}/src/app/common:ids",
     "${chip_root}/src/app/data-model-provider/tests:encode-decode",
     "${chip_root}/src/app/server-cluster",
+    "${chip_root}/src/app/server-cluster:registry",
     "${chip_root}/src/app/server-cluster/testing",
     "${chip_root}/src/lib/core:string-builder-adapters",
   ]
