+++ src/app/server-cluster/tests/TestServerClusterInterfaceRegistry.cpp
@@ -22,7 +22,7 @@
 #include <app/server-cluster/DefaultServerCluster.h>
 #include <app/server-cluster/ServerClusterContext.h>
 #include <app/server-cluster/testing/TestServerClusterContext.h>
-#include <data-model-providers/codegen/ServerClusterInterfaceRegistry.h>
+#include <app/server-cluster/ServerClusterInterfaceRegistry.h>
 #include <lib/core/CHIPError.h>
 #include <lib/core/DataModelTypes.h>
 #include <lib/core/StringBuilderAdapters.h>
