+++ src/data-model-providers/codegen/CodegenDataModelProvider.h
@@ -24,7 +24,7 @@
 #include <app/data-model-provider/ActionReturnStatus.h>
 #include <app/data-model-provider/MetadataTypes.h>
 #include <app/util/af-types.h>
-#include <data-model-providers/codegen/ServerClusterInterfaceRegistry.h>
+#include <app/server-cluster/ServerClusterInterfaceRegistry.h>
 #include <lib/core/CHIPPersistentStorageDelegate.h>
 #include <lib/support/ReadOnlyBuffer.h>
