+++ src/data-model-providers/codegen/model.gni
@@ -38,9 +38,9 @@ codegen_data_model_PUBLIC_DEPS = [
   "${chip_root}/src/app/common:attribute-type",
   "${chip_root}/src/app/data-model-provider",
   "${chip_root}/src/app:required-privileges",
-  "${chip_root}/src/app/server-cluster",
   "${chip_root}/src/data-model-providers/codegen:instance-header",
-  "${chip_root}/src/data-model-providers/codegen:registry",
+  "${chip_root}/src/app/server-cluster",
+  "${chip_root}/src/app/server-cluster:registry",
   "${chip_root}/src/app/persistence",
   "${chip_root}/src/app/persistence:default",
   "${chip_root}/src/app/persistence:singleton",
