+++ src/data-model-providers/codegen/tests/BUILD.gn
@@ -54,7 +54,6 @@ chip_test_suite("tests") {
   test_sources = [
     "TestCodegenModelViaMocks.cpp",
     "TestEmberAttributeDataBuffer.cpp",
-    "TestServerClusterInterfaceRegistry.cpp",
   ]
 
   cflags = [ "-Wconversion" ]
